feat: add JobCheck ValidateHealth support
Adds `JobCheck` to `ValidateHealth`: a Job, referenced from a ConfigMap/Secret via `JobRef`,
usable alongside the existing Lua Script and CEL evaluation checks.

Bumps libsveltos to pick up ValidateHealth.JobCheck: a Job, referenced via
JobRef from a ConfigMap/Secret, that can be used as a health check alongside
the existing Lua Script and CEL evaluation options.

Removes this repo's dependency on the private `sveltos-enterprise` module entirely.
`go.mod` no longer references it, under any build tag.
