--- Dockerfile.enterprise
@@ -1,52 +1,0 @@
-# syntax=docker/dockerfile:1.26
-
-# Build the manager binary (Enterprise build - links in the private sveltos-enterprise module)
-FROM golang:1.26.5@sha256:3aff6657219a4d9c14e27fb1d8976c49c29fddb70ba835014f477e1c70636647 AS builder
-
-ARG BUILDOS
-ARG TARGETARCH
-
-WORKDIR /workspace
-# Copy the Go Modules manifests
-COPY go.mod go.mod
-COPY go.sum go.sum
-
-# github.com/projectsveltos/sveltos-enterprise is a private module. Unlike the default
-# Dockerfile, this build always needs it (it's the enterprise variant), so it's fetched here
-# via git over SSH, authenticated with a key forwarded from the host through BuildKit's ssh
-# mount (see Makefile's docker-buildx target: --ssh default=<path-to-key>).
-RUN git config --global url."git@github.com:".insteadOf "https://github.com/" && \
-    mkdir -p -m 0700 /root/.ssh && ssh-keyscan github.com >> /root/.ssh/known_hosts
-RUN --mount=type=ssh GOPRIVATE=github.com/projectsveltos/sveltos-enterprise go mod download
-
-# Copy the go source
-COPY cmd/main.go cmd/main.go
-COPY api/ api/
-COPY lib/ lib/
-COPY controllers/ controllers/
-COPY pkg/ pkg/
-COPY internal/ internal/
-
-# Build
-RUN CGO_ENABLED=0 GOOS=$BUILDOS GOARCH=$TARGETARCH go build -tags enterprise -a -o manager cmd/main.go
-
-# Use distroless as minimal base image to package the manager binary
-# Refer to https://github.com/GoogleContainerTools/distroless for more details
-FROM gcr.io/distroless/static:nonroot
-
-ARG GIT_VERSION=unknown
-
-LABEL org.opencontainers.image.source="https://github.com/projectsveltos/addon-controller" \
-      org.opencontainers.image.url="https://projectsveltos.io" \
-      org.opencontainers.image.licenses="Apache-2.0" \
-      org.opencontainers.image.vendor="projectsveltos" \
-      org.opencontainers.image.title="addon-controller" \
-      org.opencontainers.image.description="Deploys Kubernetes add-ons and applications (Helm charts, Kustomize, raw YAML) across fleets of clusters, with built-in multi-tenancy support." \
-      org.opencontainers.image.version="$GIT_VERSION" \
-      org.opencontainers.image.revision="$GIT_VERSION"
-
-WORKDIR /
-COPY --from=builder /workspace/manager .
-USER 65532:65532
-
-ENTRYPOINT ["/manager"]
