+++ lib/crd/clustersummaries.go
@@ -1348,7 +1348,7 @@ spec:
                           type: array
                         featureID:
                           description: |-
-                            FeatureID is an indentifier of the feature (Helm/Kustomize/Resources)
+                            FeatureID is an identifier of the feature (Helm/Kustomize/Resources)
                             This field indicates when to run this check.
                             For instance:
                             - if set to Helm this check will be run after all helm
@@ -1366,6 +1366,46 @@ spec:
                             Group of the resource to fetch in the managed Cluster.
                             Required when Kind is set. Leave empty for metric-only checks.
                           type: string
+                        jobCheck:
+                          description: |-
+                            JobCheck runs a Job in the managed cluster and uses its Complete/Failed
+                            outcome as the check result. Mutually exclusive with Script and EvaluateCEL.
+                          properties:
+                            jobRef:
+                              description: |-
+                                JobRef references the Secret/ConfigMap containing the Job manifest to
+                                deploy in the managed Cluster as this check.
+                              properties:
+                                kind:
+                                  description: 'Kind of the resource. Supported kinds
+                                    are: Secrets and ConfigMaps.'
+                                  enum:
+                                  - Secret
+                                  - ConfigMap
+                                  type: string
+                                name:
+                                  description: Name of the referenced resource.
+                                  minLength: 1
+                                  type: string
+                                namespace:
+                                  description: |-
+                                    Namespace of the referenced resource.
+                                    Namespace can be left empty. In such a case, namespace will
+                                    be implicit set to cluster's namespace.
+                                  type: string
+                              required:
+                              - kind
+                              - name
+                              - namespace
+                              type: object
+                            timeout:
+                              description: |-
+                                Timeout is how long to wait for the Job to reach Complete or Failed
+                                before treating the check as failed.
+                              type: string
+                          required:
+                          - jobRef
+                          type: object
                         kind:
                           description: |-
                             Kind of the resource to fetch in the managed Cluster.
@@ -1480,6 +1520,9 @@ spec:
                       - featureID
                       - name
                       type: object
+                      x-kubernetes-validations:
+                      - message: jobCheck cannot be set together with script or evaluateCEL
+                        rule: '!has(self.jobCheck) || (!has(self.script) && !has(self.evaluateCEL))'
                     type: array
                     x-kubernetes-list-type: atomic
                   preDeleteChecks:
@@ -1519,7 +1562,7 @@ spec:
                           type: array
                         featureID:
                           description: |-
-                            FeatureID is an indentifier of the feature (Helm/Kustomize/Resources)
+                            FeatureID is an identifier of the feature (Helm/Kustomize/Resources)
                             This field indicates when to run this check.
                             For instance:
                             - if set to Helm this check will be run after all helm
@@ -1537,6 +1580,46 @@ spec:
                             Group of the resource to fetch in the managed Cluster.
                             Required when Kind is set. Leave empty for metric-only checks.
                           type: string
+                        jobCheck:
+                          description: |-
+                            JobCheck runs a Job in the managed cluster and uses its Complete/Failed
+                            outcome as the check result. Mutually exclusive with Script and EvaluateCEL.
+                          properties:
+                            jobRef:
+                              description: |-
+                                JobRef references the Secret/ConfigMap containing the Job manifest to
+                                deploy in the managed Cluster as this check.
+                              properties:
+                                kind:
+                                  description: 'Kind of the resource. Supported kinds
+                                    are: Secrets and ConfigMaps.'
+                                  enum:
+                                  - Secret
+                                  - ConfigMap
+                                  type: string
+                                name:
+                                  description: Name of the referenced resource.
+                                  minLength: 1
+                                  type: string
+                                namespace:
+                                  description: |-
+                                    Namespace of the referenced resource.
+                                    Namespace can be left empty. In such a case, namespace will
+                                    be implicit set to cluster's namespace.
+                                  type: string
+                              required:
+                              - kind
+                              - name
+                              - namespace
+                              type: object
+                            timeout:
+                              description: |-
+                                Timeout is how long to wait for the Job to reach Complete or Failed
+                                before treating the check as failed.
+                              type: string
+                          required:
+                          - jobRef
+                          type: object
                         kind:
                           description: |-
                             Kind of the resource to fetch in the managed Cluster.
@@ -1651,6 +1734,9 @@ spec:
                       - featureID
                       - name
                       type: object
+                      x-kubernetes-validations:
+                      - message: jobCheck cannot be set together with script or evaluateCEL
+                        rule: '!has(self.jobCheck) || (!has(self.script) && !has(self.evaluateCEL))'
                     type: array
                     x-kubernetes-list-type: atomic
                   preDeployChecks:
@@ -1690,7 +1776,7 @@ spec:
                           type: array
                         featureID:
                           description: |-
-                            FeatureID is an indentifier of the feature (Helm/Kustomize/Resources)
+                            FeatureID is an identifier of the feature (Helm/Kustomize/Resources)
                             This field indicates when to run this check.
                             For instance:
                             - if set to Helm this check will be run after all helm
@@ -1708,6 +1794,46 @@ spec:
                             Group of the resource to fetch in the managed Cluster.
                             Required when Kind is set. Leave empty for metric-only checks.
                           type: string
+                        jobCheck:
+                          description: |-
+                            JobCheck runs a Job in the managed cluster and uses its Complete/Failed
+                            outcome as the check result. Mutually exclusive with Script and EvaluateCEL.
+                          properties:
+                            jobRef:
+                              description: |-
+                                JobRef references the Secret/ConfigMap containing the Job manifest to
+                                deploy in the managed Cluster as this check.
+                              properties:
+                                kind:
+                                  description: 'Kind of the resource. Supported kinds
+                                    are: Secrets and ConfigMaps.'
+                                  enum:
+                                  - Secret
+                                  - ConfigMap
+                                  type: string
+                                name:
+                                  description: Name of the referenced resource.
+                                  minLength: 1
+                                  type: string
+                                namespace:
+                                  description: |-
+                                    Namespace of the referenced resource.
+                                    Namespace can be left empty. In such a case, namespace will
+                                    be implicit set to cluster's namespace.
+                                  type: string
+                              required:
+                              - kind
+                              - name
+                              - namespace
+                              type: object
+                            timeout:
+                              description: |-
+                                Timeout is how long to wait for the Job to reach Complete or Failed
+                                before treating the check as failed.
+                              type: string
+                          required:
+                          - jobRef
+                          type: object
                         kind:
                           description: |-
                             Kind of the resource to fetch in the managed Cluster.
@@ -1822,6 +1948,9 @@ spec:
                       - featureID
                       - name
                       type: object
+                      x-kubernetes-validations:
+                      - message: jobCheck cannot be set together with script or evaluateCEL
+                        rule: '!has(self.jobCheck) || (!has(self.script) && !has(self.evaluateCEL))'
                     type: array
                     x-kubernetes-list-type: atomic
                   reloader:
@@ -2017,7 +2146,7 @@ spec:
                           type: array
                         featureID:
                           description: |-
-                            FeatureID is an indentifier of the feature (Helm/Kustomize/Resources)
+                            FeatureID is an identifier of the feature (Helm/Kustomize/Resources)
                             This field indicates when to run this check.
                             For instance:
                             - if set to Helm this check will be run after all helm
@@ -2035,6 +2164,46 @@ spec:
                             Group of the resource to fetch in the managed Cluster.
                             Required when Kind is set. Leave empty for metric-only checks.
                           type: string
+                        jobCheck:
+                          description: |-
+                            JobCheck runs a Job in the managed cluster and uses its Complete/Failed
+                            outcome as the check result. Mutually exclusive with Script and EvaluateCEL.
+                          properties:
+                            jobRef:
+                              description: |-
+                                JobRef references the Secret/ConfigMap containing the Job manifest to
+                                deploy in the managed Cluster as this check.
+                              properties:
+                                kind:
+                                  description: 'Kind of the resource. Supported kinds
+                                    are: Secrets and ConfigMaps.'
+                                  enum:
+                                  - Secret
+                                  - ConfigMap
+                                  type: string
+                                name:
+                                  description: Name of the referenced resource.
+                                  minLength: 1
+                                  type: string
+                                namespace:
+                                  description: |-
+                                    Namespace of the referenced resource.
+                                    Namespace can be left empty. In such a case, namespace will
+                                    be implicit set to cluster's namespace.
+                                  type: string
+                              required:
+                              - kind
+                              - name
+                              - namespace
+                              type: object
+                            timeout:
+                              description: |-
+                                Timeout is how long to wait for the Job to reach Complete or Failed
+                                before treating the check as failed.
+                              type: string
+                          required:
+                          - jobRef
+                          type: object
                         kind:
                           description: |-
                             Kind of the resource to fetch in the managed Cluster.
@@ -2149,6 +2318,9 @@ spec:
                       - featureID
                       - name
                       type: object
+                      x-kubernetes-validations:
+                      - message: jobCheck cannot be set together with script or evaluateCEL
+                        rule: '!has(self.jobCheck) || (!has(self.script) && !has(self.evaluateCEL))'
                     type: array
                     x-kubernetes-list-type: atomic
                 type: object
@@ -2183,7 +2355,7 @@ spec:
                         type: string
                       type: array
                     featureID:
-                      description: FeatureID is an indentifier of the feature whose
+                      description: FeatureID is an identifier of the feature whose
                         status is reported
                       enum:
                       - Resources
