+++ lib/crd/profiles.go
@@ -1308,7 +1308,7 @@ spec:
                       type: array
                     featureID:
                       description: |-
-                        FeatureID is an indentifier of the feature (Helm/Kustomize/Resources)
+                        FeatureID is an identifier of the feature (Helm/Kustomize/Resources)
                         This field indicates when to run this check.
                         For instance:
                         - if set to Helm this check will be run after all helm
@@ -1326,6 +1326,46 @@ spec:
                         Group of the resource to fetch in the managed Cluster.
                         Required when Kind is set. Leave empty for metric-only checks.
                       type: string
+                    jobCheck:
+                      description: |-
+                        JobCheck runs a Job in the managed cluster and uses its Complete/Failed
+                        outcome as the check result. Mutually exclusive with Script and EvaluateCEL.
+                      properties:
+                        jobRef:
+                          description: |-
+                            JobRef references the Secret/ConfigMap containing the Job manifest to
+                            deploy in the managed Cluster as this check.
+                          properties:
+                            kind:
+                              description: 'Kind of the resource. Supported kinds
+                                are: Secrets and ConfigMaps.'
+                              enum:
+                              - Secret
+                              - ConfigMap
+                              type: string
+                            name:
+                              description: Name of the referenced resource.
+                              minLength: 1
+                              type: string
+                            namespace:
+                              description: |-
+                                Namespace of the referenced resource.
+                                Namespace can be left empty. In such a case, namespace will
+                                be implicit set to cluster's namespace.
+                              type: string
+                          required:
+                          - kind
+                          - name
+                          - namespace
+                          type: object
+                        timeout:
+                          description: |-
+                            Timeout is how long to wait for the Job to reach Complete or Failed
+                            before treating the check as failed.
+                          type: string
+                      required:
+                      - jobRef
+                      type: object
                     kind:
                       description: |-
                         Kind of the resource to fetch in the managed Cluster.
@@ -1440,6 +1480,9 @@ spec:
                   - featureID
                   - name
                   type: object
+                  x-kubernetes-validations:
+                  - message: jobCheck cannot be set together with script or evaluateCEL
+                    rule: '!has(self.jobCheck) || (!has(self.script) && !has(self.evaluateCEL))'
                 type: array
                 x-kubernetes-list-type: atomic
               preDeleteChecks:
@@ -1478,7 +1521,7 @@ spec:
                       type: array
                     featureID:
                       description: |-
-                        FeatureID is an indentifier of the feature (Helm/Kustomize/Resources)
+                        FeatureID is an identifier of the feature (Helm/Kustomize/Resources)
                         This field indicates when to run this check.
                         For instance:
                         - if set to Helm this check will be run after all helm
@@ -1496,6 +1539,46 @@ spec:
                         Group of the resource to fetch in the managed Cluster.
                         Required when Kind is set. Leave empty for metric-only checks.
                       type: string
+                    jobCheck:
+                      description: |-
+                        JobCheck runs a Job in the managed cluster and uses its Complete/Failed
+                        outcome as the check result. Mutually exclusive with Script and EvaluateCEL.
+                      properties:
+                        jobRef:
+                          description: |-
+                            JobRef references the Secret/ConfigMap containing the Job manifest to
+                            deploy in the managed Cluster as this check.
+                          properties:
+                            kind:
+                              description: 'Kind of the resource. Supported kinds
+                                are: Secrets and ConfigMaps.'
+                              enum:
+                              - Secret
+                              - ConfigMap
+                              type: string
+                            name:
+                              description: Name of the referenced resource.
+                              minLength: 1
+                              type: string
+                            namespace:
+                              description: |-
+                                Namespace of the referenced resource.
+                                Namespace can be left empty. In such a case, namespace will
+                                be implicit set to cluster's namespace.
+                              type: string
+                          required:
+                          - kind
+                          - name
+                          - namespace
+                          type: object
+                        timeout:
+                          description: |-
+                            Timeout is how long to wait for the Job to reach Complete or Failed
+                            before treating the check as failed.
+                          type: string
+                      required:
+                      - jobRef
+                      type: object
                     kind:
                       description: |-
                         Kind of the resource to fetch in the managed Cluster.
@@ -1610,6 +1693,9 @@ spec:
                   - featureID
                   - name
                   type: object
+                  x-kubernetes-validations:
+                  - message: jobCheck cannot be set together with script or evaluateCEL
+                    rule: '!has(self.jobCheck) || (!has(self.script) && !has(self.evaluateCEL))'
                 type: array
                 x-kubernetes-list-type: atomic
               preDeployChecks:
@@ -1648,7 +1734,7 @@ spec:
                       type: array
                     featureID:
                       description: |-
-                        FeatureID is an indentifier of the feature (Helm/Kustomize/Resources)
+                        FeatureID is an identifier of the feature (Helm/Kustomize/Resources)
                         This field indicates when to run this check.
                         For instance:
                         - if set to Helm this check will be run after all helm
@@ -1666,6 +1752,46 @@ spec:
                         Group of the resource to fetch in the managed Cluster.
                         Required when Kind is set. Leave empty for metric-only checks.
                       type: string
+                    jobCheck:
+                      description: |-
+                        JobCheck runs a Job in the managed cluster and uses its Complete/Failed
+                        outcome as the check result. Mutually exclusive with Script and EvaluateCEL.
+                      properties:
+                        jobRef:
+                          description: |-
+                            JobRef references the Secret/ConfigMap containing the Job manifest to
+                            deploy in the managed Cluster as this check.
+                          properties:
+                            kind:
+                              description: 'Kind of the resource. Supported kinds
+                                are: Secrets and ConfigMaps.'
+                              enum:
+                              - Secret
+                              - ConfigMap
+                              type: string
+                            name:
+                              description: Name of the referenced resource.
+                              minLength: 1
+                              type: string
+                            namespace:
+                              description: |-
+                                Namespace of the referenced resource.
+                                Namespace can be left empty. In such a case, namespace will
+                                be implicit set to cluster's namespace.
+                              type: string
+                          required:
+                          - kind
+                          - name
+                          - namespace
+                          type: object
+                        timeout:
+                          description: |-
+                            Timeout is how long to wait for the Job to reach Complete or Failed
+                            before treating the check as failed.
+                          type: string
+                      required:
+                      - jobRef
+                      type: object
                     kind:
                       description: |-
                         Kind of the resource to fetch in the managed Cluster.
@@ -1780,6 +1906,9 @@ spec:
                   - featureID
                   - name
                   type: object
+                  x-kubernetes-validations:
+                  - message: jobCheck cannot be set together with script or evaluateCEL
+                    rule: '!has(self.jobCheck) || (!has(self.script) && !has(self.evaluateCEL))'
                 type: array
                 x-kubernetes-list-type: atomic
               reloader:
@@ -1974,7 +2103,7 @@ spec:
                       type: array
                     featureID:
                       description: |-
-                        FeatureID is an indentifier of the feature (Helm/Kustomize/Resources)
+                        FeatureID is an identifier of the feature (Helm/Kustomize/Resources)
                         This field indicates when to run this check.
                         For instance:
                         - if set to Helm this check will be run after all helm
@@ -1992,6 +2121,46 @@ spec:
                         Group of the resource to fetch in the managed Cluster.
                         Required when Kind is set. Leave empty for metric-only checks.
                       type: string
+                    jobCheck:
+                      description: |-
+                        JobCheck runs a Job in the managed cluster and uses its Complete/Failed
+                        outcome as the check result. Mutually exclusive with Script and EvaluateCEL.
+                      properties:
+                        jobRef:
+                          description: |-
+                            JobRef references the Secret/ConfigMap containing the Job manifest to
+                            deploy in the managed Cluster as this check.
+                          properties:
+                            kind:
+                              description: 'Kind of the resource. Supported kinds
+                                are: Secrets and ConfigMaps.'
+                              enum:
+                              - Secret
+                              - ConfigMap
+                              type: string
+                            name:
+                              description: Name of the referenced resource.
+                              minLength: 1
+                              type: string
+                            namespace:
+                              description: |-
+                                Namespace of the referenced resource.
+                                Namespace can be left empty. In such a case, namespace will
+                                be implicit set to cluster's namespace.
+                              type: string
+                          required:
+                          - kind
+                          - name
+                          - namespace
+                          type: object
+                        timeout:
+                          description: |-
+                            Timeout is how long to wait for the Job to reach Complete or Failed
+                            before treating the check as failed.
+                          type: string
+                      required:
+                      - jobRef
+                      type: object
                     kind:
                       description: |-
                         Kind of the resource to fetch in the managed Cluster.
@@ -2106,6 +2275,9 @@ spec:
                   - featureID
                   - name
                   type: object
+                  x-kubernetes-validations:
+                  - message: jobCheck cannot be set together with script or evaluateCEL
+                    rule: '!has(self.jobCheck) || (!has(self.script) && !has(self.evaluateCEL))'
                 type: array
                 x-kubernetes-list-type: atomic
             type: object
